+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Current
 - Only one Django.js test suite
 - Each framework is tested against its own test suite
 - Make jQuery support optionnal into ``JsTestCase``
+- Improved JsTestCase output
 
 
 0.3.2 (2012-11-10)